moved navguards to separate file
